About me
Just a normal teen wanting to create amazing things for people to enjoy! I am an expert when it comes to movies and tv shows. Video games too. I am a Weeb, but taking a break from anime for now. Anyways, that's all for now. Oh, I also have a youtube channel that I plan on putting more of my time into. It is called: Okira Yoshi.